refactor: auth schema 파일들 위치 변경 (#47)
- auth 관련 schema 파일들을 shared에서 features/auth로 이동
- 관련 import 경로들 수정
