+++ src/features/auth/api/schemas/auth/index.ts
@@ -1,16 +1,16 @@
 export {
   EmailLoginRequestSchema,
   EmailLoginResponseSchema,
-} from '@/shared/api/schemas/auth/login'
+} from '@/features/auth/api/schemas/auth/login'
 export {
   TokenRefreshRequestSchema,
   TokenRefreshResponseSchema,
-} from '@/shared/api/schemas/auth/refresh'
+} from '@/features/auth/api/schemas/auth/refresh'
 export {
   LogoutRequestSchema,
   LogoutResponseSchema,
-} from '@/shared/api/schemas/auth/logout'
+} from '@/features/auth/api/schemas/auth/logout'
 export {
   UserSchema,
   TokenResponseSchema,
-} from '@/shared/api/schemas/auth/common'
+} from '@/features/auth/api/schemas/auth/common'
